fix: moved tag related components into ui-modules to avoid duplication
