--- frontend/plugins/frontline_ui/src/modules/ticket/components/tags/components/TagsGroupsAddButtons.tsx
@@ -1,14 +1,0 @@
-import { Button } from 'erxes-ui';
-import { useTagContext } from '@/ticket/components/tags/providers/TagProvider';
-
-export const TagsGroupsAddButtons = () => {
-  const { startAddingTag, startAddingGroup } = useTagContext();
-  return (
-    <div className="flex gap-2">
-      <Button onClick={startAddingGroup} variant="outline">
-        Add Group
-      </Button>
-      <Button onClick={startAddingTag}>Add Tag</Button>
-    </div>
-  );
-};
